simplify cursor ring draw state
Removed old code and state that used to allow for the cursor
ring to be different if a synthetic link was pressed, state
that allowed the ring to animate, and state that attempted to
show the ring until the link it represented was replaced by
a new page.

The new code
- shows the cursor ring indefinitely when the trackball or dpad is moved
- shows the cursor for 500 ms when the trackball or dpad is clicked
  or if the screen is tapped.

This is separate from logic that attempts to select whether to
show the cursor ring at all; this merely chooses how long to
show the cursor ring when the cursor changes.

Requires a companion change in frameworks/base

Change-Id: I309abe346f6b67ed3b665aaa79c367f4599bf1cd
http://b/2135321