Initial autofill changes in libwebcore.
Initial checkin of WebKit source for enabling AutoFill on
Android. This code calls into the chromium library to perform the
autofill magic.

There's still lots to do, but this and a corresponding change in
external/chromium enable the feature (basically) end to end.
The feature is disabled by default until we implement it more fully
with some proper UI.

To turn it on, set the ENABLE_AUTOFILL environment variable to "true"
and rebuild. This is only useful for preliminary testing though as
for now we use a precanned profile, which won't be relevant unless
your name happens to be John Smith.

Change-Id: I4a3e0e840617f7cf8f522af33ae1be560768a6c5