Merge WebKit at r67178 : Fix AndroidAnimation.
TimingFunction was previously a simple struct with an enum to
select different behaviors. It's now an abstract class with
a different subclass for each behavior. Fixing AndroidAnimation
so that it holds a RefPtr to the timing function rather than
trying to store its value directly.

See http://trac.webkit.org/changeset/67032

Change-Id: Icb7f2911aea341975531c95594ab1c30ac48cd87