clip img elements the same as text
The nav cache builds the cursor ring by collecting the bounds of
the text and images contained by the clickable node. The clip
described by the parent node is available when the bounds are
collected.

The current code clips the text. Generalize this to clip the
images also.

Also, print whether the node is transparent in the node dumper.

Change-Id: I44d6cbacb95211f191cf11b6abd5273e0712930f
http://b/2463829