Fix a bug that would cause the AutoFill server to return an error.
We were pushing the <option> elements associated with <select> tags
into the form field vector that AutoFill uses. This is not correct
as they are not used by the server and result in duplicate field
signatures in the XML requests. As the signatures should be unique,
the server returns an error response.

Change-Id: I359a952cf219e088c2cae886907f563d86756688