Remove superfluous Geolocation::stop() method rogue LayoutTests
The stop() method and LayoutTests were added in
https://android-git.corp.google.com/g/#change,53849 which
cherry-picked a WebKit change to attempt to stop leaks. This change
has been rolled back in WebKit pending a better solution, but was
retained in Android as a temporary fix.

Later changes to Geolocation in WebKit have now fixed the underlying
problem and this has been confirmed on Android.

The temporary fix was mostly removed in the merge to WebKit r65615 in
https://android-git.corp.google.com/g/#change,62318 This patch simply
removes the superfluous stop() method and LayoutTests, which were
missed in the above patch.

Bug: 3001031
Change-Id: Id17117f285cb1a8fa600945cc6a87fc336709cab