Do not merge: fix array overwrite crasher
If two nodes have the same coordinates, one is deleted.
If either has focus, the focus needs to be transfered
to the other, and the focus index recomputed, so the
index won't point to the wrong node.

If the two nodes are at the end of the list, the
index may point off the end of the array, subsequently
crashing on access.

This is a possible security issue.

Change-Id: I1ca934074637fbf68e40318fbc354e28c6b474ba
http://b/3043268