Don't include JavaScriptCore/config.h from WebCore/WebCorePrefix.h
This is a cherry-pick of the following upstream CL:
https://bugs.webkit.org/show_bug.cgi?id=47179

This violates the rules at the top of the file, and means that it
will be included before any other includes for all compilation
units.

We are now just including JavaScriptCore/config.h from
WebCore/config.h so if you include one you get both.

This created a compile error in V8Counters.cpp, which is also fixed

Change-Id: I272247161877ddc680b3aab509bc413ab3445219