Sync autofill profile data from java.
Add the JNI bindings so that we can get the AutoFill profile data
that the user entered in the Browser UI and inject that data into
the WebAutoFill object. This eliminates the John Smith profile!

Note that additional profile data such as addresses and phone
numbers will be added very shortly.

Requires a corresponding change to frameworks/base,
see https://android-git.corp.google.com/g/#change,72204

Change-Id: Ie32100123ef0d868e1f7efb5604c6fac48725967