Fix a WebCore/WebKit layering violation
Move the call to WebViewCore::updateViewport() from HTMLBodyElement
and HTMLMetaElement to the PlatformBridge so that WebCore code does
not depend on WebKit code to compile.

In addition to being a layering violation, the inclusion of
WebViewCore.h includes WebCoreJni.h, which in a future change
causes LOG to get #undefined when using the chromium net stack.

Change-Id: I1a9fd0b545080c9c0d9d26a3e5eb3b986c3986eb