Fix WebCore/WebKit layering violation.
Move the call to WebViewCore::updateTextfield() to the PlatformBridge
so that WebCore code does not depend on WebKit code to compile.

In addition to being a layering violation, the inclusion of
WebViewCore.h includes WebCoreJni.h, which in a future change
causes LOG to get #undefined when using the chromium net stack.

Change-Id: Id88d332624323e03ce82120ef55db200cbabe63b