Fixed browser crash when plugins remove themselves from the DOM.
If a plugin removes itself from the DOM and it is the only plugin
of that type on the page then the browser will attempt to unload
the plugin's shared library. This causes a problem since there is
plugin code still on the stack. The solution is to post a message
that removes the plugin's libary at a later time. This solution
was disabled in the past due to an unknown issue, but was tested
as working on both the Adobe plugin and a development sample plugin.

Change-Id: If60394336790f833d6b9333424836cbd32dc31e7
http://b/3084460