Cherry-pick WebKit security fix (webkit.org r68096) Do not merge
See http://trac.webkit.org/changeset/68096

Note that this is not an exact cherry pick of the above webkit
CL as it is dependant on http://trac.webkit.org/changeset/61094
that we do not have and provides the Element::isIdAttributName and
Element::getIdAttribute methods. 61094 is a large change so we do
not take it, rather use methods that achieve the same effect.

Change-Id: Ifae90116f794673a610dd7988ba366a0b51888d9