allow contained draws to identify a link
The first cut of rewriting hiding and clipping links if
they are obscured by a subsequent draw looks for the
contained text to find the original link in the picture.

Some links contain only bitmaps, and others, like text
fields, may contain nothing at all. To identify these,
look for bitmaps, and to cover the text field case, look for
a background erase. These draws can be used to identify
where the link is in the picture.

Because bitmaps and background erases can also obscure
links, distinguish between draws that are contained by
the link from those that overlap it.

bug:3120589
Change-Id: Ied8fe2378f27253a162f5f9636f5ade12ce6013c