Notify WebKit of redirects in the Chrome HTTP stack.
In this CL we just defer the redirect until WebKit has had a chance
to check it, and potentially cancel it.

There's a comment in the old code asking what to do if we're given
a bad request. Looks like this should never happen; it's a DCHECK in
Chrome so I've turned it into an ASSERT here.

We only follow redirects if WebKit does not modify the URL (this is
the same behaviour as Chrome).

Change-Id: I0c8b8cd61c501527a29dda5aca521a7df4a3ccef