Adjust layers when parent layer scrolls
When a page has layers, it causes the main body of the
page to become a layer as well. The scroll position
of the WebView must be applied to the layer to
translate the cursor rings correctly for hit-testing.

Today, applying the scroll position of the WebView
screws up overflow scrolling. Until this gets figured
out, skip using the scroll position in CacheFrame
adjust() and unadjust().

The position of the layer, and all of its parent layers,
should be considered when adjusting the rectangle
bounds contained by that layer. Before, only the
child layer's position was considered.

The clip used by scrollable divs may be initialized
but never set. If it is empty, ignore it, rather than
clipping out all content.

In CacheBuilder, remove a couple of unnecessary
lines -- the CacheInput initialization clears all
members.

In SelectText, reverse the order of the xfer mode
and paint objects to remove a Skia ref count assert.

In CachedInput, make the debug printout current.

Overall, added more debugging output (turned
off by default).

bug:3030370

Change-Id: Ic19c24b3bf33d081a1d0c1f8c06601dcb56ae881