Unref the picture instead of deleting it
This fixes a couple of bugs related to layers
and hardware acceleration.

The picture used by select text is now accessed
asynchronously from the UI thread. The UI thread
now unrefs it rather than deleting it outright.

The layer position may not be up to date when
the cursor rings are drawn, so update them if
the cursor ring is in a layer.

bug:3112657
bug:3030370
Change-Id: I99d07175ec30dcba895a26cf1d44bc0364430793