update dom version on style change
The DOM version number is used to know when the
nav cache is out of date. Webkit changes the version
when DOM attributes change, but not when the CSS
style property changes.

Tracking the style version fixes Google properties like
the 'Options' menu that is available on mobile
devices after a search.

Additionally, fix a crash if the root layer doesn't
exist. And, rebuild the nav cache always on touch.

bug:2628448
Change-Id: I50aa258c1b057ac7deed00f6eca37c0ee323efcf