Merge WebKit at r72274: Implement PlatformBridge memory management methods for WebKit change 71639
Get device class specific memory constraints for V8.

V8 can be more efficient in terms of memory management but to do
so needs to know about what resources are available to it.
This varies device to device on Android so add implement the functions
that the the PlatformBridge needs to query the relevant values from
Java.

MemoryUsage.cpp will compile for JSC, but needs tuning to report
correct JSC stats. It's currently a non-issue, as this routine is only
used with V8, at present.

Requires a frameworks/base change:
https://android-git.corp.google.com/g/#change,78690

Bug: 3075565

Requires a previously upstreamed webkit change:
http://trac.webkit.org/changeset/71639

Change-Id: I25ff54b5c8e25f5de5380e0d43e56551587adc5d