Ratchet back style changes that rebuild nav caches
Many style changes do not visibly affect the way
links are detected -- so limit the rebuild to
style changes that either move the size the style,
or changes that edit the style's display property.

Styles appearing or disappearing (i.e., one is 0)
is not treated as a style version change since
these events accompany dom edits which are detected
outside of this change.

bug:3181683
Change-Id: Ie62ef524a956fcd9298ce5878b632645a22e461a