use layer id to maintain selection
The pictures used to draw the page content can change
in response to refreshes, as the selection is moved or
as the screen is scrolled or zoomed.

Track the layer id, instead of the picture pointer,
to find the picture that contains the current selection.

Before, the selection would disappear as the screen
was pinched, or the highlight might disappear even
though the anchors draw, or the highlight and anchors
might be drawn offset from the screen data. This change
eliminates this class of bugs.

Change-Id: Ifc440b8aa0c9d6d232e298e42c4c544a41629442
http://b/3183492