Track all tiles in the viewport instead of only the upper left tile.
We currently store only the coordinates of the upper left tile and
leave it to the caller to compute the actual number of tiles in the
viewport. This CL computes that information upfront and then passes
a rect containing the tiles in the viewport to the necessary parties.

This also allows us to intelligently prepare tiles that are currently
off-screen but given the user's scroll direction will be on-screen
shortly.

bug: 3190926

Change-Id: I50a5922ed8bc7e8fd8e4cacd4986a9f275cd37b0