Don't set error description strings in ResourceError.
These strings may be displayed to the user, so they need to be
localized. In a separate change (I7dec8dff) I updated the Java
framework to use a standard localized message if the description
is empty, so here we just use the empty string (same as Chrome).

Bug: 3172265
Test: Load http://google.com:81 to force a timeout error, check
that the correct message is displayed.

Change-Id: I9042a2ee99cd78619cc0392cdc2cbb82ebd51743