improve left edge tap detection
Sites like Gizmodo.com layout paragraphs with extra space
between lines. Increase the slop allowed when looking for
the left edge of a paragraph to accommodate this.

Add #define to disable sideways scrolling if the line is
already entirely visible.

Make various slop consts more clearly named and commented.

A companion fix in frameworks/base is not required, but
makes scrolling as a result of taps more reliable.

bug:1331125
bug:3099569
Change-Id: I0ed89c1f53de99831e835e417b5409c1176fc841