Merge autofill files after last Chromium merge
Merge by hand the files that have been gotked implementing
AutoFill on Android. This classes use WebCore types rather
than the Chrome WebKit API types that the code would
normally use.

This brings the files up to date with the last external/chromium
merge to r63472 and updates their usages in line with that so
that the feature continues to work as intended.

Change-Id: Ic110873dc48c59555f0db76e7ad0c138c2fe96f2