By default, provide native AutoFill with an empty profile.
If no profile is set then AutoFill will not inform Java that it has
found a form that could be autofilled. We want to be able to offer
the user the option Java side to set up a profile in this case, so
we set by default an empty profile that is active until the user
syncs their real profile. This ensures that the AutoFill detection
code will not early out.

Change-Id: I474b25c5b31dae59e4ddb23c7f398d013e9d3beb