Add maybeSavePassword() method to WebFrame.
This calls through to the new entrypoint on the Java size
(see change I3f3f84a1). I've done some modest refactoring
to minimize code duplication. The way we obtain a pointer
to the frame is still a bit messy, but I couldn't find a
really clean solution to that.

Change-Id: Ia5c12f615b8c41f1c846bbf6fd6dcfe56f3e2a51