reenable draw extras when GL is turned on
This captures the drawing in the UI thread, then passes
the drawing to the tile imaging thread.

The draw extras interface now takes an additional rectangle,
and each draw extra fills in the inval area formed as that
part is drawn.

The old extra implementation in GLWebViewState has been
removed. The inval portion of the setBaseLayer call has been split
out so it can be called directly.

bug:3161294
Change-Id: I28d3e6879059770b973e7c0f7c0796909f7359aa