Set useCachedCredentials to true on first auth request.
This flag previously defaulted to false, but the Android CTS
specifies that a "sufficient" implementation should return true
on the first authentication request. (The second and subsequent
auth requests aren't tested.)

This CL sets the flag to true on the first authentication request
for a given WebRequest, but false for subsequent requests (i.e.,
if the first attempt to authenticate failed).

Using a counter rather than a flag in WebRequest for readability.
"isFirstAuthRequest" would be a bit ambiguous as a member variable,
but "authRequestCount" seems pretty clear.

Change-Id: Id9baae4c30111f743e079f90940078b07498b168