Fix crash in JavaClassV8
It is possible that the m_name member of JavaClass is uninitialized as
we possibly early out from the constructor if we could not load the Java
class from the JVM. In this case when we call the dtor we try to free
uninitialized memory and crash. m_name is not actually used anywhere
so we just remove it.

Will upstream to webkit.

Bug: 3222518
Change-Id: I9354e9914157d2c4afd72dcce9122329c2486c89