use style ignoring pending style sheets
If the style is not yet computed, attempting to
compute the style may blow up (e.g., the font points
to zero and is dereferenced). The webkit way around
this appears to be to ignore pending style sheets,
even though six lines down from this change, the
regular webkit makes the original styleForElement
call.

This does fix sfgate.com, so it's not a bad thing to
do, and the real style change should retrigger this
code path later on.

bug:3194102
Change-Id: Ia558e3d49c94ed817d69d528b9e914f906f62a66