WtfStringToJString should be able to return a valid string on empty input
It is sometimes desirable to return a valid but 0-length string created
from a 0-length WTF string but currently the WtfStringToJString function
will always return a null jstring on an empty input. Add a parameter to
specify that it should return an empty jstring on empty WTF string and
default it to false.

Change-Id: Ia449f1f9ae9b37c3da7d0272424e4fd4878ceb31