Fix crash when syncing autofill settings.
Since we moved the autofill initiation code out of the constructor
and into it's own method, we may crash if there is an autofill profile
set but the feature is turned off (i.e. the user had it on at one point
and had a profile set up but turned it off without deleting the
profile). That's because in this case we execute setAutoFillProfile()
in WebSettings without having first called the new init method. It makes
sense to not sync any profiles though when the feature is not enabled,
so add that check.

Change-Id: I9d99e96347770e5b08c982513f2370cb770cbe94