Refactor AutoFill so it runs lazily.
This CL refactors the AutoFill code so that we only search for forms
when we need to, i.e. the user has focused a form field. This helps
page load times. We also re-search the document for forms if the dom
version number has changed. These changes fix a bug where a crash
occurs if the form is focused before the page has finished loading.

We also move the callsite of formFieldFocused so that it only occurs
when the IME is requested which is a more appropriate place called less
frequently than it's current callsite in EditorClientAndroid.

Bug: 3243567
Change-Id: I04fe13944d3f8248197d71a0e58c21f01c4abdd0