keep rects separate across canvas push
The nav cache identifies visible elements by accumulating
the drawing commands used in an interesting part of the picture,
then eliminating the portions that are too large or overdrawn.

Consecutive draws of the same type are combined, so that, for
instance, glyphs assimilate to form a string, or bitmaps form
a nine-patch. Draws of the same type shouldn't be combined when
they represent different objects.

Break apart draws of the same object if there is a push/pop
pair surrounding the previous draw of that object type.

bug:3241830
bug:3241992
Change-Id: I4eb65a0ae0d923d4950cda88e16cb4bbacfde663