When the tile get painted, it should be cleared first.
This will clear the residue for the areas not in viewport, either
the area belongs to the tiles inside or outside the viewport.
In another word, area in the viewport will be redraw anyway,
but area not in the viewport will get an invalidate only, that
is why we need to clear the content when invalidating.

bug: 3253934
Change-Id: I29b88c20cb5e68d6c0f917e7fa9fb22b395d6321