call hideCursor directly, avoiding a newer cache
The WebView interface hideCursor gets a newer cache if
one is available. This interface was also called when
the cursor animation completes, before the cursor is drawn.

Getting a newer cache at this time can cause the data
in m_ring to be invalid.

Pass the root cache stored as part of the ring data instead,
so that the ring is drawn consistently.

bug:3262275
Change-Id: Id205a10f96329e9d5ac794a8537f25a05918cafd