WebResponse headers should be case-insensitive.
The CTS test WebViewTest.testSetDownloadListener was failing due to
a disagreement over the capitalization of "Content-Disposition". Rather
than making our stack exactly match the CTS server, it looks like the
headers should actually be case-insensitive. (Note that we still preserve
the case of the headers exactly as reported by the server, so hopefully
this won't break anything else.)

BUG=3242063
TEST=WebViewTest.testSetDownloadListener

Change-Id: I0931da98aab9080127e355e51d89ac3fde12e7c2