improve text selection in scrollable divs
Detect which layer is hit by looking for scrollable layers.
In this case it isn't necessary to see what is drawn or where,
because the layer has already been prescreened by the parent clip.

Adjust the x/y hit by the layer's position. This corrects the
position when the layer scrolls.

Fix CachedLayer's debug output to be in sync with LayerAndroid.h

SelectText edge cases are corrected when there is not text to hit.

To find the closest character, first look at the edge of the
text line. If the hit is inside the line of text, measure from the
center of the character. This helps when textlines overlap.

bug:3275625
bug:3191699

Change-Id: I924cc0432f6dace32e0a7444fe69ef15dca2e77f