use same matrix left when reading picture
In select text, the nav cache picture is read multiple times.

If the matrix is different, it's possible that the text
metrics will be off by one when the text position is near
to but not equal to an integer (e.g., 315.999969). Pinning
the matrix to zero makes the matrices the same in the bug case.

Use the simpler doIRect interface if the glyph rec isn't needed.

Check for contentRender() before dereferencing it.

bug:3315620
Change-Id: I4bfaf9f9b7bf12f66b29f4609d744bceaf5b2964