Merge WebKit at r72805: Cherry-pick WebKit change 73062
This change is the upstreaming of local changes in
WebCore/platform/network. It did not generate conflicts in the merge to
r73109 as both upstream and local versions matched at r73109. But since
it appears after r72805, it was inadvertantly removed as part of the
initial merge by Git to r72805.

See http://trac.webkit.org/changeset/73062

Note that this is a backwards merge from Chromium release 9.0.600.0
to 9.0.597.0, to align with the Chromium 9 stable release branch.

Change-Id: I6a266e1270cbc1f155065ae1c56c491ccf33e497