Merge WebKit at r72805: Cherry-pick WebKit change 72948
This fixes CodeGeneratorV8.pm

It is a re-commit of 'Merge WebKit at r72274: Fix CodeGeneratorV8.pm',
which was inadvertantly removed as part of the initial merge by Git to
r72805.

See http://trac.webkit.org/changeset/72948

Note that this is a backwards merge from Chromium release 9.0.600.0
to 9.0.597.0, to align with the Chromium 9 stable release branch.

Change-Id: Iaa56a5a503aaaf4d996ebc44d66d557d12838cfa