return the scrollable layer's content bounds
Compute and return the scrollable layer's bounds in
content space and return it to the java side.

The layer's bounds is used to determine if the
select text touch point is close to the edge and
should invoke scrolling.

Also, added a null check to selectBestAt

Requires a companion change in frameworks/base

bug:3191699
bug:3334943

Change-Id: Id066997f56da59f37350c14a4cc43fe4a7ce7312