Merge Chromium at 9.0.597.55: Move formsSeen call to IO thread.
When we search the document for forms to be autofilled, we make a URL
request to the AutoFill server to ask it for form field data. With
Chromium r67655, this URL request needs to be made on a Chrome thread.

This is a simple fix for the merge, which blocks until the call
completes.

See http://src.chromium.org/viewvc/chrome?view=rev&revision=67655

Change-Id: Iad96314e4ecb66ee7e65c1ef846709e9c41b0e2e