Fix some sites not responding.
Enter compositing for iframes after all layout tasks have completed.  Remember
the state of overflow scroll similar to RenderLayer so that it doesn't change
while updating the compositing state.

Return false in requiresAcceleratedCompositing if the iframe is hidden.  I verified
that this fails on Safari and have submitted a bug.

Bug: 3325187
Change-Id: I20a9e1a595d14518f4c7a3e01fc1c4fa9035262c