don't draw buttons while they are created
There are three threads all hammering at the same resource.
The WebViewCore thread records the buttons on the page.
The WebView thread rerecords the button pictures to show selection and focus.
The GLWebViewState thread draws the picture containing the button pictures
into the tile.

Prevent drawing the picture while the buttons are recorded
by sharing the button mutext with the GL tile creation.

bug:3354678
Change-Id: I245ed47ad5a6d1fe28ea870bd7c557937eed1164