Fix hit testing inside layers.
A new webkit merge added a couple methods to ClipRects that were not copying the
hit test rect.  This make the clip rect empty during hit testing which was
clipping out all nodes in scrollable layers.

When tracking a layer, use the foreground layer if present as it will contain
the right nodes.

When looking for a scrollable layer, traverse the children in reverse drawing
order to find the top-most visible layer.  This allows orkut.com to scroll.

Remove a couple of casts and headers.  Use relaxAdoptionRequirement() to avoid a
RefCounted assert.  Swap the texture owner hash set during deletion as release
can modify the iterator concurrently.

Update the scrollbars (even though we don't have any) when scrolling a layer to
keep the scroll position of the scrollbar in sync with the layer.

Rewrite a little bit of PluginPackageAndroid.  m_module may be non-null during
load.  If it is null, load the library and store it in m_module.  Follow the
regular path assuming m_module is not null.

Bug: 3373179
Change-Id: If07ec9735b30c1e98e363667378a8d253a841a45