Fixed element positioning fix
This is for fixed bottom/ right, in the zooming case.
The width and height in webkit should be calculated as the visibleRect
we use for drawing. And that is document coordinate for visible size.

Basically it is send the Rect info from webView to webViewCore.
Then at RenderBox, it will go through the PlatformBridge to pick the
info up.

Notice that the touch is not 100% working yet, the layout call in webkit
can have a early return such that fixed element layer didn't get update.
And a touch/click is not really causing the whole layout update yet.
That will be addressed in seperate change though.

bug:3404129
Change-Id: I225d41815143a05d540ed32bfc76f823603ca89c