Get layers showing again
Lazily get the maximum texture size from GL when
we first need it. This avoids calling GL functions
before GL has been set up and ensures it is done on
the correct (ie. UI) thread.

Also move the tiles expansion setting from WC thread
to UI thread as TilesManager is not thread safe when
creating the singleton instance.

This makes change 4596782e unnecessary so we remove it.

Bug: 3412928

Change-Id: I740974d8cc0ef4f66956cd6a07b058bfa7ca767f