A couple of fixes for scrolling layers.
Update the layer structure during compositing sync. Make sure to turn off
scrolling if it changes.

Since I replace the content layer with a different structure, append the
children to the foreground layer so they move with the content. Remove the
clipping layer in the compositor for scrollable layers.

For children of scrollable layers, add back in the content scroll offset so that
layers are positioned relative to 0,0.

Bug: 3416512
Change-Id: I74de90aac2dcf67bd969d0b85f440343123ecab9