fix buttons in GL
Buttons failed to draw when selected with the keyboard, and
sometimes when selected by touch, for three reasons:

1) GL short-circuits if the last extra rectangle and the
   curent extra rectangle match. When buttons change color,
   the rectangles do match, but the picture still needs to
   be redrawn.

2) Buttons do not have cursor rings, so the button's bounds
   was not passed to GL.

3) The timeout logic to switch from a focused state to a normal
   state was incorrect.

This adds a flag to GLWebViewState::setExtra that can skip the
equal rectangle case. Note that the equal rectangle test hides
a more serious bug where the setExtra is called repeatedly -- so
it would be worthwhile to examine this further.

The logic that determines if the cursor is a button was split
out so that the cursor ring can be associated with a button even
if the cursor is hidden.

bug:3364248
Change-Id: I75944cb045486c128743aa13922d20758ccf783f