Do not merge: Cherry-pick Flash performance CL from Master
Bug: 3471589

Fix some of the performance issues when using the flash plugin embedded.

What happened was that MediaLayer was always telling us to repaint the
screen; we could repaint the same frame a dozen of times unecessarily.
This in itself was wasteful but should have been ok, but the stream
of commands plus the compositing caused the GPU commands used by flash
to stall until they were executed...
The compounded impact drasticaly lowered the performances.

This is not a full fix as those conditions (us repainting the screen
a lot) happens anyway when zooming or scrolling; but this should improve
performances in the general case.

Change-Id: I79a75c759fd5968cb822616eba0caa8c77e75835