fix select text space detection
Some helpers like isSpace() are used in two contexts: while the
picture is parsed, and after the best answer in the picture has
been found. The proper paint must be used in each case.

Associate paints with candidate and best glyphs, so that the
appropriate paint can be used as needed.

bug:3470564
Change-Id: I28bcc82dbe727ee3ef0716e7657e61a982467a40