Possible fix for bug 3410366 SEGV in when hitting "go back" on an ssl cert warning
RefCounting the handle that is sent to webkit (and java). Note
that if java does not call back with either cancel or proceed it
will leak. The documentation for in WebViewClient.java says:

Notify the host application to handle a ssl certificate error request
(display the error to the user and ask whether to proceed or not). The
host application has to call either handler.cancel() or handler.proceed()
as the connection is suspended and waiting for the response. The default
behavior is to cancel the load.

This means it is clearly an error not to call any of them, and leaking
in that case is OK.

Change-Id: Idac713833395edfbec3ae9b5612a34c2cf25f4e7