Fix Browser ANR
The dtor of GLWebViewState is calling the dtor of its TiledPage
variables; the dtor of TiledPage wait until any pending painting
operations running in the TextureGenerator thread are done. The painting
operations ultimately paint using the current GLWebViewState's base
layer. We were doing unref() on the baselayer before destroying the
TiledPage...

bug:3429924
Change-Id: I685e7fca5e5bae796f808debbf7e4ce668b83e6b